@vanja-p 2025-04-08
This is failing a test added in https://github.com/buildbuddy-io/buildbuddy/pull/4626. Maybe we don't want to rely on `digest.size_bytes` being correct?

Also, I'm not sure if these code paths are also used by the action cache. If they are, then using `digest.size_bytes` like this is definitely not correct.

If either of these is a problem, I do have a version which is just a bit slower, but returns all the read bytes, even if there are more than `digest.size_bytes`.